% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap_info_interval.R
\name{bootstrap_info_interval}
\alias{bootstrap_info_interval}
\title{Bootstrap resample an information measure}
\usage{
bootstrap_info_interval(
  a_vec,
  b_vec,
  measure_func = associationsubgraphs::calc_mutual_info,
  B = 500,
  ...
)
}
\arguments{
\item{a_vec}{an \href{https://adv-r.hadley.nz/vectors-chap.html}{atomic vector} of observations (can be logical, factor, character, or integer).}

\item{b_vec}{paired observations for second variable.}

\item{measure_func}{Any function that takes as its first two arguments to vectors of categorical values and returns a number can be used as a "measure".}

\item{B}{Number of bootstrap resamples to gather}

\item{...}{Any additional arguments to \code{measure_func} outside of \code{a_vec} and \code{b_vec}.}
}
\value{
An array of values returned from running \code{measure_func} on bootstrap resampled observations from provided data.
}
\description{
Runs a simple bootstrap resampling of your observations on information measure of choice to assess uncertainty.
}
\examples{

N <- 100
pois_mi_samples <- bootstrap_info_interval(
  a_vec = rpois(N, 5),
  b_vec = rpois(N, 8),
  B = 150
)
hist(pois_mi_samples)

# Any function that takes as its first two arguments to vectors of categorical
# values and returns a number can be used as a "measure".
pois_je_samples <- bootstrap_info_interval(
  a_vec = rpois(N, 5),
  b_vec = rpois(N, 8),
  measure = calc_joint_ent,
  B = 150
)
hist(pois_je_samples)

# You can pass any arguments to the entropy function you may desire as well...
binom_mi_samples <-
  bootstrap_info_interval(
    a_vec = rbinom(N, size = 1, p = 0.1),
    b_vec = rbinom(N, size = 1, p = 0.08),
    B = 150,
    measure = function(a, b, scalar) calc_joint_ent(a, b)*scalar,
    scalar = -2
  )
hist(binom_mi_samples)
#'
}
