% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{build_count_vec}
\alias{build_count_vec}
\title{Build counts observation vec}
\usage{
build_count_vec(n, fill_indices, fill_values)
}
\arguments{
\item{n}{Total number of observations (aka the length of the final vector).}

\item{fill_indices}{Integer vector of all indices an occurrence happened.}

\item{fill_values}{Array of same length as \code{fill_indices} that contain the
integer counts for the filled observations.}
}
\value{
Array of length \code{n} with every observation indicated by \code{fill_indices} filled with the matching entry in \code{fill_values} and \code{0} elsewhere.
}
\description{
Build counts observation vec
}
\examples{
build_count_vec(10, c(1,3,4), fill_values = c(10, 20, 30))
}
