% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relative_association_functions.R
\name{build_relative_associations}
\alias{build_relative_associations}
\title{Generate a relative association network}
\usage{
build_relative_associations(
  association_pairs,
  strength_col = "strength",
  rank_based = FALSE,
  impute_missing
)
}
\arguments{
\item{association_pairs}{dataframe with columns \code{a} and \code{b} representing the
ids of the variables or nodes and columns \code{strength} that is a numeric
indicator of strength of association (higher = stronger).}

\item{rank_based}{Should the relative strength be calculated using ranks?
Here instead of using the raw values of strength, the ranking of edges is
used with the max "strength" being the number of edges. Set to \code{FALSE} to
use the raw strengths.}

\item{impute_missing}{If not all possible pairs are present in
\code{association_pairs}, this parameter controls how the function deals with
these missing values. Current options are \code{"minimum"}(lowest seen edge
value is substituted) and \code{zero} (\code{0} is substituted). See the section
"missing association pairs" for more details.}
}
\value{
A new association pairs dataframe (with columns \code{a}, \code{b}, and
\code{strength}) where \code{strength} is transformed to relative strength.
}
\description{
Takes as input a raw association network and normalizes it to
relative-associations. Every edge's strength is transformed by dividing by
the average strength of each of the variables it connects. E.g. if edge \code{a-b}
has strength \code{1} and \code{a}'s average association is \code{0.5} and \code{b}'s average
association is \code{0.75} that means this edge has a higher than normal weight
for both nodes and its normalized strength is \code{1/((0.5 + 0.75)/2) = 1.6}.
}
\section{missing association pairs}{


If there are missing pairs - I.e. not all \code{n*(n-1)/2} combinations of
variables are represented - the function will throw a warning fill in the
missing values with the minimum seen association strength.

The following options are available using the \code{impute_missing} argument:
\itemize{
\item \code{"minimum"}: The lowest seen strenght is assumed to be the value for all missing pairs (same
as default but no warning is given).
\item \code{"zero"}: Value of \code{0} is substituted. This option only makes sense if your association is
naturally lower bounded at \code{0}.
\item \code{"ignore"}: Here no action is taken and the averages are just used as is. Since it is likely this missingness is not at random it's likely the best option is an imputation of some sort.
}
}

\examples{

}
