% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_ent_from_table.R
\name{calc_ent_from_table}
\alias{calc_ent_from_table}
\title{Calculate entropy of a contingency table of one or two variables.}
\usage{
calc_ent_from_table(t, N = sum(t), log_base = 2)
}
\arguments{
\item{t}{Contingency table as produced by \code{base::table()}}

\item{N}{Total number of observations in table. Used to convert counts to
empirical probabilities.}

\item{log_base}{Base of logarithm used. \code{2} will return entropy in bits.}
}
\value{
Entropy value for variable(s) making up table.
}
\description{
Calculate entropy of a contingency table of one or two variables.
}
\seealso{
base::table
}
