% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_entropy_net.R
\name{calc_entropy_net}
\alias{calc_entropy_net}
\title{Title}
\usage{
calc_entropy_net(
  pairs,
  id_col,
  target_col,
  info_func = calc_mutual_info,
  possible_targets,
  parallel = TRUE,
  subset_pairs = FALSE
)
}
\arguments{
\item{pairs}{Dataframe with at least two columns: \code{id} and \code{target}. \code{id}
corresponds to the variables to be associated with each other and \code{target}
corresponds to the observations they had. This is in essence a tidy version
of an tabular dataset with columns as \code{id} and rows as \code{target}. If data is
count based an addition \code{count} column can be included. (Not yet
implemented).}

\item{id_col}{Name of column containing variable id (I.e. nodes of the
association network).}

\item{target_col}{Name of the column containing the target for occurance
(I.e. the data used to infer association between nodes).}

\item{info_func}{What information function should be used? Like all of the
\verb{calc_*} functions in this package this takes as its arguments two paired
vectors of observations as input and returns a number corresponding to
strength of association.}

\item{possible_targets}{If there are some \code{target} values that are never seen
in your dataset then they can be provided here with the form of a
characterr, factor, or integer vector of all target values.}

\item{parallel}{Should the processing be run in parallel? Three options. A
logical \code{TRUE} will run in parallel over every core available using
\code{future::multisession}. A number will run in parallel over that many cores.
E.g. \code{parallel = 5} will run in parallel on \code{5} core. Last, for the most
control a \code{future} processing plan can be passed (I.e. what is run by
\code{future::plan()} to set up your parallel processing environment.) This
allows you to run on clusters etc..}

\item{subset_pairs}{Should the algorithm be run on a random subset of
possible combinations? Useful for making sure everything is good before
running larger jobs. Set to the number of pairs you want to calculate
strength for.}
}
\value{

}
\description{
Title
}
