% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/info_functions.R
\name{calc_joint_ent}
\alias{calc_joint_ent}
\title{Calculate joint entropy of two vectors}
\usage{
calc_joint_ent(a_vec, b_vec, log_base = 2)
}
\arguments{
\item{a_vec}{an \href{https://adv-r.hadley.nz/vectors-chap.html}{atomic vector} of observations (can be logical, factor, character, or integer).}

\item{b_vec}{paired observations for second variable.}

\item{log_base}{Base of logarithm used. \code{2} will return information measures in bits.}
}
\value{

}
\description{
Calculate joint entropy of two vectors
}
\examples{


}
