% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subgraphs_functions.R
\name{calculate_subgraph_structure}
\alias{calculate_subgraph_structure}
\title{Find all subgraphs in pairs for every subset of edges}
\usage{
calculate_subgraph_structure(association_pairs, strength_column = "strength")
}
\arguments{
\item{association_pairs}{dataframe with columns \code{a} and \code{b} representing the
ids of the variables or nodes and columns \code{strength} that is a numeric
indicator of strength of association (higher = stronger).}
}
\value{
Dataframe with the following columns for each unique subgraph state.\tabular{lll}{
   \strong{Column} \tab  \tab \strong{Description} \cr
   \code{step} \tab  \tab Integer step, aka the number of unique edge strengths in network at state \cr
   \code{n_edges} \tab  \tab How many edges have been added thus far \cr
   \code{strength} \tab  \tab The lowest strength of the edge(s) added \cr
   \code{n_nodes_seen} \tab  \tab How many unique nodes/variables are currently in network \cr
   \code{n_subgraphs} \tab  \tab How many isolated subgraphs/components of size > 2 are in current network \cr
   \code{n_triples} \tab  \tab How many isolated subgraphs of size 3 or larger are in current network \cr
   \code{max_size} \tab  \tab Size in number of nodes of the largest current subgraph \cr
   \code{max_rel_size} \tab  \tab Proportion of all seen nodes (\code{n_nodes_seen}) that the largest subgraph includes. Large values indicate presence of a giant-component. \cr
   \code{avg_size} \tab  \tab Average size of subgraphs \cr
   \code{avg_density} \tab  \tab Average density of subgraphs. Scale from >0 - 1. Where a density of 1 is a fully-connected subgraph. \cr
   \code{subgraphs} \tab  \tab List column of summary stats for each subgraph at a given step. See the \code{subgraph} list column section for more info. \cr
}
}
\description{
Given a dataframe of edges with strength between nodes this function returns
info on every subgraph state achieved by adding nodes in one-at-a-time in
descending order of strength.
}
\section{\code{subgraph} list column}{


The subgraph list column in the results contains information on the present subgraphs at each step. It is a list with the following format, but can be turned into a dataframe/tibble easily with \code{dplyr::as_tibble/as.data.frame}.\tabular{lll}{
   \strong{Column} \tab  \tab \strong{Description} \cr
   \code{id} \tab  \tab Integer ID for subgraph. Can be used to track subgraph evolution over steps. \cr
   \code{size} \tab  \tab How many variables/nodes subgraph has \cr
   \code{density} \tab  \tab Density of subgraph. Scale from >0 - 1. Where a density of 1 is a fully-connected subgraph. \cr
   \code{strength} \tab  \tab How many unique nodes/variables are currently in network \cr
   \code{first_edge} \tab  \tab 0-based integer index of first edge that made up subgraph. Used internally to match subgraphs in interactive visualizations with these results. \cr
}
}

\examples{

virus_net \%>\%
  dplyr::arrange(dplyr::desc(strength)) \%>\%
    head(1000) \%>\%
    calculate_subgraph_structure()

}
