% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_null_info_binary.R
\name{draw_null_info_binary}
\alias{draw_null_info_binary}
\title{Sample from null distribution for mutual information between binary occurrence observations}
\usage{
draw_null_info_binary(N, Na, Nb, log_base = 2)
}
\arguments{
\item{N}{Total number of observations}

\item{Na}{Number of times the a-variable occurs}

\item{Nb}{Number of times the b-variable occurs}

\item{log_base}{Base of logarithm used. \code{2} will return entropy in bits.}
}
\value{
Single row dataframe with settings for draw and mutual information (\code{mi}) for current draw.
}
\description{
Generates an example realization from the null distribution (two arrays are independent) of mutual information between two vectors of paired binary occurances.
}
\examples{

draw_null_info_binary(10, 5, 2)
}
