% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{expand_combinations}
\alias{expand_combinations}
\title{Expand all combinations}
\usage{
expand_combinations(n, self_pairs = FALSE)
}
\arguments{
\item{n}{number of unique items}

\item{self_pairs}{Should self-pairs be included? I.e. 2-2 and 4-4?}
}
\value{
dataframe with two columns \code{a_index} and \code{b_index} that contain all pairs.
}
\description{
Expands all possible pairs of integers (\code{n*(n - 1)/2}) (order doesn't matter)
for \code{1:n}. Useful for creating combinations of elements based on their
indices in an array.
}
\examples{

# Default options dont have self-pairs
expand_combinations(4)

# Self pairs can be added, though!
expand_combinations(4, self_pairs = TRUE)
}
