% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subgraphs_functions.R
\name{explore_component_structure}
\alias{explore_component_structure}
\title{Find all components in pairs for every subset of edges (c++ version)}
\usage{
explore_component_structure(association_pairs, strength_column = "strength")
}
\arguments{
\item{association_pairs}{dataframe with columns \code{a} and \code{b} representing the
ids of the variables or nodes and columns \code{strength} that is a numeric
indicator of strength of association (higher = stronger).}
}
\value{
Dataframe with the following columns for the subgraph info
}
\description{
Given a dataframe of edges with strength between nodes this function returns
info on every subgraph state achieved by adding nodes in one-at-a-time in
descending order of strength.
}
\examples{

virus_net \%>\%
  dplyr::arrange(dplyr::desc(strength)) \%>\%
    head(1000) \%>\%
    explore_component_structure()

}
