% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{find_components}
\alias{find_components}
\title{Find all components in pairs for every subset of edges (c++ version)}
\usage{
find_components(associations, a_col = "a", b_col = "b", w_col = "w")
}
\arguments{
\item{associations}{Dataframe of association between two ids with a strength}

\item{a_col, b_col}{Names of columns that store the id's for the association pair}

\item{w_col}{Name of the column storing the strength of association}
}
\description{
Given a dataframe of edges with strength between nodes this function returns
info on every component state achieved by adding nodes in one-at-a-time in
descending order of strength.
}
