% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_subgraphs.R
\name{find_subgraphs}
\alias{find_subgraphs}
\title{Find subgraphs in pairs}
\usage{
find_subgraphs(association_pairs)
}
\arguments{
\item{association_pairs}{dataframe with columns \code{a} and \code{b} representing the
ids of the variables or nodes and columns \code{strength} that is a numeric
indicator of strength of association (higher = stronger).}
}
\value{
Dataframe with the node \code{id}, that nodes \code{degree} in the pair edges
provided and the \code{subgraph_id} corresponding to the isolated subgraph the
node belongs to.
}
\description{
Given a dataframe of edges between nodes this function returns a dataframe of
each unique node with a subgraph id attached indicating its shared membership
in an isolated subgraph with other nodes.
}
\examples{
virus_net \%>\%
  dplyr::arrange(dplyr::desc(strength)) \%>\%
    head(200) \%>\%
    find_subgraphs()

}
