% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relative_association_functions.R
\name{gather_avg_strength}
\alias{gather_avg_strength}
\title{Gather average association strength}
\usage{
gather_avg_strength(association_pairs)
}
\arguments{
\item{association_pairs}{dataframe with columns \code{a} and \code{b} representing the
ids of the variables or nodes and columns \code{strength} that is a numeric
indicator of strength of association (higher = stronger).}
}
\value{
Dataframe with \code{id} and \code{avg_strength} columns for each unique variable in association pairs
}
\description{
Takes an association pairs dataframe and collapses it to compute each
variable/ids average strength. Used for normalizing pairs or investigating
patterns in average strength.
}
\examples{
gather_avg_strength(virus_net)

}
