% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subgraphs_functions.R
\name{visualize_all_subgraphs}
\alias{visualize_all_subgraphs}
\title{Visualize all possible subgraphs}
\usage{
visualize_all_subgraphs(all_subgraphs)
}
\arguments{
\item{all_subgraphs}{results dataframe from calling \code{\link{find_all_subgraphs}}}
}
\description{
Companion function to \code{\link{find_all_subgraphs}} to visualize results.
Shows a streamgraph of subgraph's by size, total number of subgraphs, and
mergers of subgraphs that took place at every unique strength cutoff for
association network.
}
\examples{

virus_net \%>\%
  dplyr::arrange(desc(strength)) \%>\%
    head(500) \%>\%
    find_all_subgraphs() \%>\%
    visualize_all_subgraphs()

}
