% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize_association_network.R
\name{visualize_subgraph_structure}
\alias{visualize_subgraph_structure}
\title{Interactive association subgraphs visualization}
\usage{
visualize_subgraph_structure(
  association_pairs,
  node_info,
  subgraph_results,
  trim_subgraph_results = TRUE,
  warn_of_mismatches = TRUE
)
}
\arguments{
\item{association_pairs}{dataframe with columns \code{a} and \code{b} representing the
ids of the variables or nodes and columns \code{strength} that is a numeric
indicator of strength of association (higher = stronger).}

\item{node_info}{Optional dataframe that has a column \code{id} that corresponds
to the variables codded in \code{a} and \code{b} of \code{association_pairs} that contains
additional info nade available on hover in visualization.}

\item{subgraph_results}{Dataframe of subgraph results as returned by
\code{\link{explore_subgraph_structure}}. If it isnt provided it is
calculated. Automatic calculation will slow down code depending on how
large dataset is.}

\item{trim_subgraph_results}{Should subgraph results after a giant
subgraph has taken over be filtered out? Rule for filtering is at least
10\% of the variables are in subgraphs and largest subgraph contains less
than 95\% of all variables in subgraphs. Allows for easier investigating of
the subgraph structure over strength.}

\item{warn_of_mismatches}{If there are differences in the ids present in
\code{association_pairs} and \code{node_info} should a warning be issued?}
}
\value{
Interactive javascript visualization of association network
subgraphs at all possible cut-points
}
\description{
Produces an interactive plot of network structure for all possible subgraph
arrangements for passed association pairs with summary statistics to guide
exploration of cut-points.
}
\examples{
visualize_subgraph_structure(
  virus_net,
  node_info = virus_host_viruses \%>\%
    dplyr::rename(id = virus_id) \%>\%
    dplyr::mutate(color = ifelse(type == "RNA", "orangered", "steelblue"))
)

}
