% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{build_occurrence_vec}
\alias{build_occurrence_vec}
\title{Build logical occurrence vec}
\usage{
build_occurrence_vec(n, fill_indices)
}
\arguments{
\item{n}{Total number of observations (aka the length of the final vector).}

\item{fill_indices}{Integer vector of all indices an occurrence happened.}
}
\value{
Array of length \code{n} with \code{TRUE} in every place indicated by \code{fill_indices} and \code{FALSE} elsewhere.
}
\description{
Build logical occurrence vec
}
\examples{
build_occurrence_vec(c(1,3,5), 6)
}
