% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/info_functions.R
\name{calc_ent_from_table}
\alias{calc_ent_from_table}
\title{Calculate entropy of a contingency table of one or two variables.}
\usage{
calc_ent_from_table(t, N = sum(t), log_base = 2)
}
\arguments{
\item{t}{Contingency table as produced by \code{base::table()}}

\item{N}{Total number of observations in table. Used to convert counts to
empirical probabilities.}

\item{log_base}{Base of logarithm used. \code{2} will return information measures in bits.}
}
\value{
Entropy value for variable(s) making up table.
}
\description{
Calculate entropy of a contingency table of one or two variables.
}
\examples{
N <- 15
a_vec <- rbinom(N, 1, 0.2)
b_vec <- rbinom(N, 1, 0.8)
calc_ent_from_table(table(a_vec, b_vec), N)

# Switch base of log to switch units of entropy
calc_ent_from_table(table(a_vec, b_vec), N, log_base = exp(1))
}
\seealso{
base::table
}
