% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{calculate_subgraph_structure_rcpp}
\alias{calculate_subgraph_structure_rcpp}
\title{Find all subgraphs in pairs for every subset of edges (c++ version)}
\usage{
calculate_subgraph_structure_rcpp(
  associations,
  a_col = "a",
  b_col = "b",
  w_col = "w"
)
}
\arguments{
\item{associations}{Dataframe of association between two ids with a strength}

\item{a_col, b_col}{Names of columns that store the id's for the association pair}

\item{w_col}{Name of the column storing the strength of association}
}
\description{
Given a dataframe of edges with strength between nodes this function returns
info on every subgraph state achieved by adding nodes in one-at-a-time in
descending order of strength.
}
