% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize_association_network.R
\name{visualize_association_network}
\alias{visualize_association_network}
\title{Interactive association network visualization}
\usage{
visualize_association_network(
  association_pairs,
  node_info,
  measure_title = "association",
  alphaDecay = 0.01,
  n_neighbors = 5
)
}
\arguments{
\item{association_pairs}{dataframe with columns \code{a} and \code{b} representing the
ids of the variables or nodes and columns \code{strength} that is a numeric
indicator of strength of association (higher = stronger).}

\item{node_info}{Optional dataframe that has a column \code{id} that corresponds
to the variables codded in \code{a} and \code{b} of \code{association_pairs} that contains
additional info nade available on hover in visualization.}

\item{measure_title}{Name of measure used to quantify strength of association
between the variables.}

\item{alphaDecay}{Paramter to control how fast the force-layout converges.
Lower values mean a better but slower layout, higher means faster. See the
\href{https://github.com/d3/d3-force#simulation_alphaDecay}{d3-force docs} for
more info.}

\item{n_neighbors}{How many neighbors for a hovered node should be shown?}
}
\value{
Interactive javascript visualization of association network
}
\description{
Produces an interactive plot of nodes connected by edges scaled by the
strength of association.
}
\details{
The plot automatically detects isolated subgraphs (groups of nodes with
connections only within the group) and lays them out in a grid to avoid
overlap between unrelated parts of the network.
}
\examples{

virus_net \%>\%
  dplyr::arrange(desc(strength)) \%>\%
  head(300) \%>\%
  visualize_association_network()

}
